[refactor] GenerateGYRHueColorPalette: extract from VerticalBarChart for reuse
