+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/agile/AgileScreen.kt
@@ -36,6 +36,7 @@ import com.rwmobi.kunigami.ui.components.LoadingScreen
 import com.rwmobi.kunigami.ui.components.ScrollbarMultiplatform
 import com.rwmobi.kunigami.ui.components.koalaplot.VerticalBarChart
 import com.rwmobi.kunigami.ui.theme.getDimension
+import com.rwmobi.kunigami.ui.utils.generateGYRHueColorPalette
 import io.github.koalaplot.core.bar.DefaultVerticalBarPlotEntry
 import io.github.koalaplot.core.bar.DefaultVerticalBarPosition
 import io.github.koalaplot.core.bar.VerticalBarPlotEntry
@@ -65,6 +66,12 @@ fun AgileScreen(
 
     val dimension = LocalDensity.current.getDimension()
     val lazyListState = rememberLazyListState()
+    val colorPalette = remember {
+        generateGYRHueColorPalette(
+            saturation = 0.6f,
+            lightness = 0.6f,
+        )
+    }
 
     Box(modifier = modifier) {
         if (uiState.rates.isNotEmpty()) {
@@ -133,6 +140,7 @@ fun AgileScreen(
                                         "$timeRange\n${vatInclusivePrice}p"
                                     }
                                 },
+                                colorPalette = colorPalette,
                                 backgroundPlot = { graphScope ->
                                     graphScope.HorizontalLineAnnotation(
                                         location = 24.55,
